refactor: extract PendingTranscriptionRetryFailureHandler.swift from TranscriptionRecoveryPresenters.swift
Splits `PendingTranscriptionRetryFailureHandler` (~35 lines) out of
TranscriptionRecoveryPresenters.swift into its own file. The remaining
RetryTranscriptionStatusPresenter + RetryableSessionPreservationPresenter
stay together.

Byte-identical move of lines 61-95. Both files import Foundation.

TranscriptionRecoveryPresenters.swift: 164 → 127 lines. Tests: 667.
